Refactor image thumbnails to live in components section.
