+++ app/src/main/java/com/jordangellatly/hermosahappyhour/ui/components/ImageThumbnail.kt
@@ -1,4 +1,4 @@
-package com.jordangellatly.hermosahappyhour.ui.home
+package com.jordangellatly.hermosahappyhour.ui.components
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -11,13 +11,57 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import com.jordangellatly.hermosahappyhour.ui.components.HermosaHappyHourSurface
+import coil.compose.AsyncImage
+import coil.request.ImageRequest
+import com.jordangellatly.hermosahappyhour.R
 import com.jordangellatly.hermosahappyhour.ui.theme.HermosaHappyHourTheme
+
+@Composable
+fun RestaurantImage(
+    imageUrl: String,
+    contentDescription: String?,
+    modifier: Modifier = Modifier,
+    elevation: Dp = 0.dp
+) {
+    HermosaHappyHourSurface(
+        color = Color.LightGray,
+        elevation = elevation,
+        shape = CircleShape,
+        modifier = modifier
+    ) {
+        AsyncImage(
+            model = ImageRequest.Builder(LocalContext.current)
+                .data(imageUrl)
+                .crossfade(true)
+                .build(),
+            contentDescription = contentDescription,
+            placeholder = painterResource(R.drawable.placeholder),
+            modifier = Modifier.fillMaxSize(),
+            contentScale = ContentScale.Crop,
+        )
+    }
+}
+
+@Preview
+@Composable
+fun RestaurantImagePreview() {
+    HermosaHappyHourTheme {
+        RestaurantImage(
+            imageUrl = "\"https://scontent-lax3-2.xx.fbcdn.net/v/t39.30808-6/292336365_435159251951538_4078078271979326231_n.jpg?_nc_cat=107&ccb=1-7&_nc_sid=09cbfe&_nc_ohc=C_tAC12x4o0AX9uN-Yg&_nc_ht=scontent-lax3-2.xx&oh=00_AfAAxhYzzBx-bK8gDhRHmOeMrzUxDDsDMZeCaTyFHRr3Ug&oe=638F052F\"",
+            contentDescription = null,
+            modifier = Modifier
+                .size(120.dp)
+        )
+    }
+}
 
 @Composable
 fun DealImage(
